Separate model tests to separate test file
